Implement a context module for terraform
When creating modules we want to scope them to the environment and
application they are meant for.   This will allow us to ship the same
resources multiple times for different environments in the same AWS acc.

We also want a standard place to set tags for these resources so we can
have a mature tagging strategy.